refactor(leaderboard): split monolithic Leaderboard.jsx into 10 focused components
Extracted sub-components into dedicated files under Leaderboard/components/:
- AnimatedCounter, RankMovementIndicator, LiveStatusBadge, PodiumCard
- LeaderboardHero, LeaderboardPodium, LeaderboardCategoryFilters
- LeaderboardControls, LeaderboardStatsCards, LeaderboardTable

Main file reduced from 1103 to 377 lines.
Each component now has a single responsibility and is independently testable.
